refactor(mcp_internal_actions): consolidate server files by replacing `server.ts` with `index.ts` and adjust related imports
