+++ front/lib/actions/mcp_internal_actions/servers/confluence/index.ts
