+++ front/lib/actions/mcp_internal_actions/servers/data_warehouses/index.ts
