+++ front/lib/actions/mcp_internal_actions/servers/slack_bot/constants.ts
