+++ front/lib/actions/mcp_internal_actions/servers/slack_bot/index.ts
