+++ front/lib/actions/mcp_internal_actions/servers/tables_query/index.ts
